Conflicts:
	CMakeLists.txt
	ethminer/main.cpp
	libdevcore/CommonData.cpp
	libdevcore/CommonData.h
	libethash-cl/CLMiner.cpp
	libethash-cl/ethash_miner_kernel.cl
	libethash-cpu/CMakeLists.txt
	libethash-cpu/CPUMiner.cpp
	libethash-cpu/CPUMiner.h
	libethash-cuda/CMakeLists.txt
	libethash-cuda/CUDAMiner.cpp
	libethash-cuda/dagger_shuffled.cuh
	libethash-cuda/ethash_cuda_miner_kernel.cu
	libethash-cuda/ethash_cuda_miner_kernel.h
	libethcore/CMakeLists.txt
	libethcore/Farm.cpp
	libethcore/Miner.h
	libpoolprotocols/stratum/EthStratumClient.cpp

+++ libdevcore/CommonData.h
@@ -166,7 +166,7 @@ inline std::string toHex(u256 val, HexPrefix prefix = HexPrefix::DontAdd)
     return (prefix == HexPrefix::Add) ? "0x" + str : str;
 }
 
-inline std::string toHex(uint64_t _n, HexPrefix _prefix = HexPrefix::DontAdd, short _bytes = 16)
+inline std::string toHex(uint64_t _n, HexPrefix _prefix = HexPrefix::DontAdd, int _bytes = 16)
 {
     // sizeof returns the number of bytes (not the number of bits)
     // thus if CHAR_BIT != 8 sizeof(uint64_t) will return != 8
@@ -176,7 +176,7 @@ inline std::string toHex(uint64_t _n, HexPrefix _prefix = HexPrefix::DontAdd, sh
     return (_prefix == HexPrefix::Add) ? "0x" + ret.str() : ret.str();
 }
 
-inline std::string toHex(uint32_t _n, HexPrefix _prefix = HexPrefix::DontAdd, short _bytes = 8)
+inline std::string toHex(uint32_t _n, HexPrefix _prefix = HexPrefix::DontAdd, int _bytes = 8)
 {
     // sizeof returns the number of bytes (not the number of bits)
     // thus if CHAR_BIT != 8 sizeof(uint64_t) will return != 4
